code review changes

+++ src/main/java/uk/gov/justice/digital/hmpps/whereabouts/services/court/VideoLinkBookingService.kt
@@ -294,14 +294,13 @@ class VideoLinkBookingService(
       prisonApiService.getLocation(it) ?: throw ValidationException("$prefix locationId $it not found in NOMIS.")
     }
   }
-
+  @Transactional
   fun deleteAppointments(appointmentId: Long) {
-    val videoLinkAppointments = videoLinkAppointmentRepository.findAllByAppointmentId(appointmentId)
-    videoLinkAppointments.forEach { videoLinkAppointment ->
-      when (videoLinkAppointment.hearingType == MAIN) {
-        true -> videoLinkBookingRepository.delete(videoLinkAppointment.videoLinkBooking)
-        false -> videoLinkAppointmentRepository.delete(videoLinkAppointment)
-      }
+    val videoLinkAppointment = videoLinkAppointmentRepository.findOneByAppointmentId(appointmentId)
+
+    when (videoLinkAppointment != null && videoLinkAppointment.hearingType == MAIN) {
+      true -> videoLinkBookingRepository.delete(videoLinkAppointment.videoLinkBooking)
+      else -> videoLinkAppointmentRepository.delete(videoLinkAppointment)
     }
   }
 }
